backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-genai>=0.2.0
requests==2.31.0
flask-sqlalchemy==3.1.1
flask-bcrypt==1.0.1
pyjwt==2.8.0

